+++ src/main/java/com/company/AlaaAdinWebsite/service/imp/ClientServiceImp.java
@@ -31,7 +31,7 @@ public class ClientServiceImp implements ClientService {
     }
 
     @Override
-    public void deleteById(int theId) {
+    public void deleteById(Long theId) {
 
     }
 }
